Move common iniMetadata functions into separate file
